pyproject.toml
[project]
name = "open-market-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.44",
    "minio>=7.2.20",
    "playwright>=1.58.0",
    "polars>=1.38.0",
    "python-dotenv>=1.2.1",
]
